+++ Source/Engine/Core/ITimeSystem.h
@@ -14,6 +14,6 @@ public:
 
 	virtual ObjectStatus getStatus() = 0;
 
-	virtual const TimeData getCurrentTime(unsigned int timezone_adjustment = 8) = 0;
-	virtual const long long getCurrentTimeFromEpoch() = 0;
+	virtual const TimeData getCurrentTime(uint32_t timezone_adjustment = 8) = 0;
+	virtual const int64_t getCurrentTimeFromEpoch() = 0;
 };
