+++ Source/Engine/Core/InnoTimer.h
@@ -1,16 +1,18 @@
 #pragma once
+#include "../Common/STL14.h"
+
 enum class TimeUnit { Microsecond, Millisecond, Second, Minute, Hour, Day, Month, Year };
 
 struct Timestamp
 {
-	unsigned short Year;
-	unsigned short Month;
-	unsigned short Day;
-	unsigned short Hour;
-	unsigned short Minute;
-	unsigned short Second;
-	unsigned short Millisecond;
-	unsigned short Microsecond;
+	uint32_t Year;
+	uint32_t Month;
+	uint32_t Day;
+	uint32_t Hour;
+	uint32_t Minute;
+	uint32_t Second;
+	uint32_t Millisecond;
+	uint32_t Microsecond;
 };
 
 class InnoTimer
@@ -21,6 +23,6 @@ public:
 	static bool Tick();
 	static bool Terminate();
 
-	static const unsigned long long GetCurrentTimeFromEpoch(TimeUnit time_unit);
-	static const Timestamp GetCurrentTime(unsigned int time_zone_adjustment);
+	static const uint64_t GetCurrentTimeFromEpoch(TimeUnit time_unit);
+	static const Timestamp GetCurrentTime(uint32_t time_zone_adjustment);
 };
