+++ Source/Engine/RenderingBackend/DX11RenderingBackend/DX11RenderingBackend.cpp
@@ -67,8 +67,8 @@ bool DX11RenderingBackendNS::createPhysicalDevices()
 {
 	HRESULT result;
 
-	unsigned int numModes;
-	unsigned long long stringLength;
+	uint32_t numModes;
+	uint64_t stringLength;
 
 	// Create a DirectX graphics interface factory.
 	result = CreateDXGIFactory(__uuidof(IDXGIFactory), (void**)&g_DXRenderingBackendComponent->m_factory);
@@ -122,7 +122,7 @@ bool DX11RenderingBackendNS::createPhysicalDevices()
 	// When a match is found store the numerator and denominator of the refresh rate for that monitor.
 	auto l_screenResolution = g_pModuleManager->getRenderingFrontend()->getScreenResolution();
 
-	for (unsigned int i = 0; i < numModes; i++)
+	for (uint32_t i = 0; i < numModes; i++)
 	{
 		if (displayModeList[i].Width == l_screenResolution.x
 			&&
@@ -144,7 +144,7 @@ bool DX11RenderingBackendNS::createPhysicalDevices()
 	}
 
 	// Store the dedicated video card memory in megabytes.
-	g_DXRenderingBackendComponent->m_videoCardMemory = (int)(g_DXRenderingBackendComponent->m_adapterDesc.DedicatedVideoMemory / 1024 / 1024);
+	g_DXRenderingBackendComponent->m_videoCardMemory = (int32_t)(g_DXRenderingBackendComponent->m_adapterDesc.DedicatedVideoMemory / 1024 / 1024);
 
 	// Convert the name of the video card to a character array and store it.
 	if (wcstombs_s(&stringLength, g_DXRenderingBackendComponent->m_videoCardDescription, 128, g_DXRenderingBackendComponent->m_adapterDesc.Description, 128) != 0)
@@ -234,7 +234,7 @@ bool DX11RenderingBackendNS::createSwapChain()
 	featureLevel = D3D_FEATURE_LEVEL_11_1;
 
 	// Create the swap chain, Direct3D device, and Direct3D device context.
-	unsigned int creationFlags = 0;
+	uint32_t creationFlags = 0;
 #ifdef _DEBUG
 	creationFlags |= D3D11_CREATE_DEVICE_DEBUG;
 #endif // _DEBUG
@@ -546,7 +546,7 @@ bool DX11RenderingBackendNS::terminate()
 
 DX11MeshDataComponent* DX11RenderingBackendNS::addDX11MeshDataComponent()
 {
-	static std::atomic<unsigned int> meshCount = 0;
+	static std::atomic<uint32_t> meshCount = 0;
 	meshCount++;
 	auto l_rawPtr = g_pModuleManager->getMemorySystem()->spawnObject(m_MeshDataComponentPool, sizeof(DX11MeshDataComponent));
 	auto l_MDC = new(l_rawPtr)DX11MeshDataComponent();
@@ -557,7 +557,7 @@ DX11MeshDataComponent* DX11RenderingBackendNS::addDX11MeshDataComponent()
 
 DX11MaterialDataComponent* DX11RenderingBackendNS::addDX11MaterialDataComponent()
 {
-	static std::atomic<unsigned int> materialCount = 0;
+	static std::atomic<uint32_t> materialCount = 0;
 	materialCount++;
 	auto l_rawPtr = g_pModuleManager->getMemorySystem()->spawnObject(m_MaterialDataComponentPool, sizeof(DX11MaterialDataComponent));
 	auto l_MDC = new(l_rawPtr)DX11MaterialDataComponent();
@@ -568,7 +568,7 @@ DX11MaterialDataComponent* DX11RenderingBackendNS::addDX11MaterialDataComponent(
 
 DX11TextureDataComponent* DX11RenderingBackendNS::addDX11TextureDataComponent()
 {
-	static std::atomic<unsigned int> textureCount = 0;
+	static std::atomic<uint32_t> textureCount = 0;
 	textureCount++;
 	auto l_rawPtr = g_pModuleManager->getMemorySystem()->spawnObject(m_TextureDataComponentPool, sizeof(DX11TextureDataComponent));
 	auto l_TDC = new(l_rawPtr)DX11TextureDataComponent();
